+++ ops/cli-dep.nix
@@ -1,4 +1,4 @@
 {
-  version = "0.6.7";
-  sha256 = "1kcab39rwangb4m1viw58ppvf1ps4i75i2dgapzyqklq87f1bmpi";
+  version = "0.6.9";
+  sha256 = "0sc8mkab9g8a1gj6mq99mpj6ki54plm8902h3bamzf0v9ir6ijnh";
 }